chore(e2e): Reorganize files + fix issue in login
